심화과제 작업
