+++ src/main.tsx
@@ -1,12 +1,10 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import App from "./App";
-import { ThemeProvider } from "./ThemeProvider.tsx";
+// import { ThemeProvider } from "./ThemeProvider.tsx";
 
 createRoot(document.getElementById("root")!).render(
-	<StrictMode>
-		<ThemeProvider>
-			<App />
-		</ThemeProvider>
-	</StrictMode>,
+  <StrictMode>
+    <App />
+  </StrictMode>,
 );
